pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["StephenFeddes <feddessteve@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.28.0"
fastapi = "^0.110.0"
pydantic = "^2.6.3"
gunicorn = "^21.2.0"
apscheduler = "^3.10.4"
pymongo = {extras = ["srv"], version = "^4.6.2"}
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
numpy = "^1.26.4"
pandas = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "uvicorn app.main:app"
